issue632: move getMaximumCommonType to PrettyPrintVisitor
