+++ javaparser-core/src/main/java/com/github/javaparser/ast/nodeTypes/NodeWithVariables.java
@@ -86,54 +86,4 @@ public interface NodeWithVariables<N extends Node> {
         return type;
     }
 
-    /**
-     * Returns the type that maximum shared type between all variables.
-     * The minimum common type does never include annotations on the array level.
-     *
-     * <br/>For <code>int a;</code> this is int.
-     * <br/>For <code>int a,b,c,d;</code> this is also int.
-     * <br/>For <code>int a,b[],c;</code> this is also int.
-     * * <br/>For <code>int[] a[][],b[],c[][];</code> this is int[][].
-     */
-    default Type getMaximumCommonType() {
-        // we use a local class because we cannot use an helper static method in an interface
-        class Helper {
-            // Conceptually: given a type we start from the Element Type and get as many array levels as indicated
-            // From the implementation point of view we start from the actual type and we remove how many array
-            // levels as needed to get the target level of arrays
-            // It returns null if the type has less array levels then the desired target
-            Type toArrayLevel(Type type, int level) {
-                if (level > type.getArrayLevel()) {
-                    return null;
-                }
-                for (int i=type.getArrayLevel();i>level;i--) {
-                    type = ((ArrayType)type).getComponentType();
-                }
-                return type;
-            }
-        }
-
-        Helper helper = new Helper();
-        int level = 0;
-        boolean keepGoing = true;
-        // In practice we want to check for how many levels of arrays all the variables have the same type,
-        // including also the annotations
-        while (keepGoing) {
-            final int currentLevel = level;
-            // Now, given that equality on nodes consider the position the simplest way is to compare
-            // the pretty-printed string got for a node. We just check all them are the same and if they
-            // are we just just is not null
-            Object[] values = getVariables().stream().map(v -> {
-                Type t = helper.toArrayLevel(v.getType(), currentLevel);
-                return t == null ? null : t.toString();
-            }).distinct().toArray();
-            if (values.length == 1 && values[0] != null) {
-                level++;
-            } else {
-                keepGoing = false;
-            }
-        }
-        return helper.toArrayLevel(getVariables().get(0).getType(), --level);
-    }
-
 }
